mockQuery structure is now package of functions
